Move createOpenApiRequest out of the internal/net directory, remove exports of other internal/net code from @osdk/api
